http://bugzilla.liquidsys.com/show_bug.cgi?id=1792 Bug: 1792
Specify "tcon" attribute for conv moves and deletes, to avoid unwanted
side effects. If a user moves a conv from "foo" to "bar", we don't want
to move any msgs it has in Trash/Sent/Junk to "bar" as well. Also, if the
user is in the Trash folder and deletes a conv, we don't want to hard-delete
its msgs that are in other folders.