http://bugzilla.liquidsys.com/show_bug.cgi?id=1405 bug: 1405
Conversations now have both a normalized subject and a raw subject.  The raw subject (with all the "Re: ", etc.) is returned to the client and is the same as the subject of the earliest message in the conversation.  Thus, emptying the Trash or hard-deleting an individual message may change the Subject of a message.  SenderList was tweaked to throw a RefreshException when the first message is detected as removed.

The format of Conversation metadata has changed to be the same as all other items.  This will cause internal parse errors on load, but they'll be detected and the metadata should automatically be recalculated.