http://bugzilla.liquidsys.com/show_bug.cgi?id=2561 Bug 2561: service.FAILURE when tagging/untagging via outlook
Sam and I could not reproduce this bug.  I suspect that recent activity
in the transaction code inadvertently fixed the problem.  Added a new
TestConcurrency unit test, which runs multiple threads that perform
concurrent read and write operations in the mailbox.

Added some timing information to the test runner output in ZimbraSuite,
so that we can keep track of unit tests that take a long time.

Factored out search code in TestTags, so it can be used by other tests.