http://bugzilla.liquidsys.com/show_bug.cgi?id=3911 bug 3911
deriving max_connections from available memory turned
out to be problematic - it was resulting in too few
max_connections.  We were also not setting the JDBC
driver's maxActive.  We now set that based on
max_connections:

  max_connections = maxActive + 4 

The 4 is so that there is some slack for mysql CLI
invocations.

The minimum bar on maxActive is now 10.  Out of the box,
on systems with 4GB RAM you we will get max connections
of about 33: 10% of system memory goes to mysql per
thread buffers and each thread takes about 12MB.  
When you have less than 1GB of RAM (or so), we peg
maxActive at 10, and reduce the per thread buffers.