http://bugzilla.liquidsys.com/show_bug.cgi?id=1378 Bug: 1378
If the autocomplete widget is showing and the user clicks elsewhere,
hide it.

Also began a bunch of DWT event cleanup. This is part one.

- DwtControl._setIERolloverEventHdlrs() folded into
  DwtControl._setMouseEventHdlrs()

- Clean up DwtControl._mouseEvent(). By default, the mouse
event is set to stop propagation and return false; widget-based
listeners may override those settings. That mechanism replaces
the use of the "_populated" property.

- Outside mousedown events (used by menu and autocomplete)
are handled through the shell rather than document.body. That
means that the handler is a widget listener rather than a DOM-based
handler.

- Removed autocomplete's isShowing() method, replaced it with
call to getVisible().