Async fix for IE
IE appears to run the onreadystatechange handler of the
XML_HTTP request object even in sync mode, so we need to
set it to an empty function on sync requests. Problem was that
after an async search, that handler was set and got run,
resulting in the response being handled twice. For example,
creating a tag would then show two of the same tag.