http://bugzilla.liquidsys.com/show_bug.cgi?id=3908 Bug: 3908
Server returns error when create/modify volume requests are sent without name or path.

Also, I removed file/mbox bits parameters from SOAP interface and zmvolume utility.  For now we will stick to default values (file/mbox = 12 bits, file group/mbox group = 8 bits).  We cannot allow these settings to change in an existing volume because existing files will become inaccessible.  So they are only useful when creating a new volume, but it's very difficult to explain how the current scheme works.  We'll need to replace it with an easier-to-understand scheme that is compatible with the current one, and only then revive the setting on volume creation command.