http://bugzilla.liquidsys.com/show_bug.cgi?id=4256 Bug: 4256
Don't allow creating or modifying a volume that would result in duplicate id, name or path.  Unique index is added to name and path columns of volume table.  path column is a TEXT column and only the first 255 characters is used in the index.  That should be adequate for most paths.