http://bugzilla.liquidsys.com/show_bug.cgi?id=1624 http://bugzilla.liquidsys.com/show_bug.cgi?id=1601 bug 1624 bug 1601
Put to rest error reporting problems with SendMsg - atleast from the
server perspective.  There are now three error codes:

    SEND_ABORTED_ADDRESS_FAILURE - this is raised when (a) atleast one
    RCPT TO: addresses was rejected AND (b) we are in "don't send to
    anyone if any address fails" mode (this is the default out of the
    box setting).  Eg:

        msg - Invalid address: foo@bolton.liquidsys.com
        code - mail.SEND_ABORTED_ADDRESS_FAILURE

        msg - Invalid addresses: foo@bolton.liquidsys.com,bar@bolton.liquidsys.com
        code - mail.SEND_ABORTED_ADDRESS_FAILURE

    SEND_PARTIAL_ADDRESS_FAILURE - this is raised when message was
    successfully sent to a subset of addresses, and not sent to some
    invalid addresses.  Tim said he would one day change SengMsg so a
    demand partial send is okay - this is so you can delete
    appoinments which may include invalid addresses.

    SEND_FAILURE - the bucket for all other types of errors - such as
    message was too big.  Eg:
        msg - SMTP server reported: 552 Error: message too large
        code - mail.SEND_FAILURE

I have big stick waiting in my office in anticipation of the jokers
who are going to reply to this and say they want that postfix SMTP
chatter localized.

In the long run our web UI could be more sophisticated and highlight
addresses that failed, and in the delete appt case offer a "go ahead
and do this any way" check box, which when selected, the UI can send a
partial-send-ok flag in the SendMsg request.  Also note that, in the
normal email send case the UI should be prepared to deal with both
ABORTED and PARTIAL responses - there is an LDAP attr to allow partial
sends all the time - though I don't see why anyone would want to do
that.  I'll let QA open an RFE for the items in this paragraph if they
want.

yeah - I know - more in the check-in message than in the code.