http://bugzilla.liquidsys.com/show_bug.cgi?id=4358 Bug: 4358
Throw an exception when an update request is made to change volume type of a volume that is assigned as a current volume.  The admin is supposed to assign another volume to be current volume of that type (if the type is secondary message, it's also possible to unset current secondary volume) to make this volume non-current.  Only then can the volume's type be changed.