http://bugzilla.liquidsys.com/show_bug.cgi?id=504 bug: 504
For multipart/alternative part, scan its child parts and pick text/plain over text/* over other content types. As a result, only the most-text-like part of the multiple alternatives is indexed.

If no text/* part is found, all alternatives are indexed as before.

Note that for appointments, only the text/plain part is now indexed. The text/calendar part is ignored, which should be fine.