http://bugzilla.liquidsys.com/show_bug.cgi?id=4000 Bug: 4000
Handle timeouts on client calls.

The default timeout for a client request is 60 seconds,
after which a dialog will appear giving the user the option
to cancel the request.

The setBusy stuff in DwtShell now takes an ID, since we
may have multiple outstanding async requests. We need to
be able to keep the timers for the requests separate.