http://bugzilla.liquidsys.com/show_bug.cgi?id=504 bug: 504
Back out this change until Kirk or I get a chance to fix it -- this was completley breaking calendaring.  This would cause the server to not analyze text/calendar parts if the message also included a text/html or a text/plain part underneath a multipart/alternative.

The standard message structure for an iTIP calendaring message is:

multipart/alternative
   text/plain
   [text/html]
   text/calendar

...if we find a text/plain part, we cannot decide to not analyze the text/calendar part: we might want to skip various other text/* parts, but we definitely don't want to skip the text/calendar part.