http://bugzilla.liquidsys.com/show_bug.cgi?id=4459 bug: 4459
Move all bulk writes outside the mailbox-synchronized block in the non-NIO ImapHandler.  There are still a few writes inside the synchronzied block, but none more than a few bytes.  Fingers crossed and shit.

No change to the NIO stuff, as they will write to a buffer and flush from a separate writer thread.