http://bugzilla.liquidsys.com/show_bug.cgi?id=4511 bug: 4511
Make Mailbox.getEffectivePermissions callable on *any* item, not just folders.  To facilitate that, add MailItem.checkRights (it was just on Folder) and move the folder-and-tags logic out of MailItem.canAccess into there.

Also, make a copy of the listener list and iterate over *that* during notifications, just in case one of the listeners decides to dequeue itself during notification.