http://bugzilla.liquidsys.com/show_bug.cgi?id=4431 Bug: 4431
Don't require -t <volume type> option in -sc (set current volume) operation.  A GetVolume request is made to find out the volume type, then current volume of that type is set to the volume ID provided.

No change is made to the SOAP API, and therefore Admin Console is not affected.