http://bugzilla.liquidsys.com/show_bug.cgi?id=4839 bug: 4839
Catch Throwable when decoding folder names that might be IMAP UTF-7, as it seems to be throwing an exception (actually an Error) that shouldn't be thrown.  When such a Throwable is caught, use the raw folder name instead of UTF-7 decoding it.  (We used to return BAD, but while that's technically correct it's probably not for the best.)