http://bugzilla.liquidsys.com/show_bug.cgi?id=4838 Bug: 4838
Fix the exception handling that resulted in the message
"Permission denied..." being displayed vertically.

The exception can be generated by calling the XMLHttpRequest's
open() method with a bad URL in AjxRpcRequest.invoke(). [thanks
Kevin H] Due to some incorrect exception handling in the async code,
the exception became a string by the time it got to _handleException().

Fixed the exception handling in ZmCsfeCommand. We didn't see this one
much because the exception has to happen before the request is sent.