Stat logging update.  Removed StopWatch calls from SOAP document handlers, since we're timing SOAP requests in SoapServlet.  Added timing for LDAP lookups and the database connection pool.
There's still a little strangeness with the stat code.
Currently Accumulators are initialized lazily.  If an
Accumulator is initialized after stats have already been
logged, we start logging columns that weren't there in
previous lines.  I'll fix this once I think of a clever
way to do the initialization without introducing nasty
interdependencies.

Added a tweak to Mailbox.getConfig(), so that it sets
success=true when nothing failed.