http://bugzilla.zimbra.com/show_bug.cgi?id=5001 Bug: 5001
Don't throw NullPointerException when "-" is sent as id in GetFreeBusy request.  It is treated the same way as any other invalid string, by returning a successful response showing the free/busy info is unknown.  From the response it is not possible to distinguish between a valid address/id for which free/busy info is not known, and an invalid string such as "-" or "&&&".