http://bugzilla.zimbra.com/show_bug.cgi?id=1279 http://bugzilla.zimbra.com/show_bug.cgi?id=1611 http://bugzilla.zimbra.com/show_bug.cgi?id=1757 http://bugzilla.zimbra.com/show_bug.cgi?id=2421 http://bugzilla.zimbra.com/show_bug.cgi?id=2544 http://bugzilla.zimbra.com/show_bug.cgi?id=2639 http://bugzilla.zimbra.com/show_bug.cgi?id=2723 http://bugzilla.zimbra.com/show_bug.cgi?id=3062 http://bugzilla.zimbra.com/show_bug.cgi?id=3140 http://bugzilla.zimbra.com/show_bug.cgi?id=3244 http://bugzilla.zimbra.com/show_bug.cgi?id=3271 http://bugzilla.zimbra.com/show_bug.cgi?id=3600 Bug: 1279, 1611, 1757, 2421, 2544, 2639, 2723, 3062, 3140, 3244, 3271, 3600
Rewrite filters UI.

This is pretty much a complete rewrite of the code that
underlies the filter rules UI. The details view for adding
or editing a filter rule is now a dialog, rather than a view
that shows up in the list. There is now a filter rules
controller to manage the view and handle events. Changes
to the filter rules are made immediately - there is no need to
hit the "Save" button.

TODO
- Clean up dialog for IE
- Quick add