DEATH TO FF_HACK!!!!
Added style.overflow:auto to the base div of DwtBaseDialog and set this._ffHackDisabled = true. This means that any input fields in the dialog will now show their cursor and we don'tneed the firefox hack anymore for dialogs. So Dialogs should be much faster on FF. 

You may now all be seated!

P.S. I hope this really works ;-)