Integrate change 19453 from CRAY to Main:
----------
Change 19453 by tim@TIMBRE on 2006/01/11 19:32:25

Fix "no such item" exception being thrown during search.

I finally got enough info from dogfood to repro this one, it turns out to be a longstanding problem with the ItemId/IndexId splitup: the search code was using the *indexId* to generate the Hit object if the result happened to be in an object with multiple mime parts (ie an attachment)....even so, this usually doesn't get noticed, UNLESS you happen to have deleted (and emptied) the original item from your mailbox.

Affected files ...

... //depot/CRAY/ZimbraServer/src/java/com/zimbra/cs/index/AppointmentHit.java#3 edit
... //depot/CRAY/ZimbraServer/src/java/com/zimbra/cs/index/ConversationHit.java#3 edit
... //depot/CRAY/ZimbraServer/src/java/com/zimbra/cs/index/DBQueryOperation.java#8 edit
... //depot/CRAY/ZimbraServer/src/java/com/zimbra/cs/index/MessageHit.java#3 edit
... //depot/CRAY/ZimbraServer/src/java/com/zimbra/cs/index/MessagePartHit.java#3 edit
... //depot/CRAY/ZimbraServer/src/java/com/zimbra/cs/index/NoteHit.java#3 edit
... //depot/CRAY/ZimbraServer/src/java/com/zimbra/cs/index/ZimbraQueryResultsImpl.java#5 edit