http://bugzilla.zimbra.com/show_bug.cgi?id=5576 Integrate change 20406 from CRAY to Main:
----------
Change 20406 by tim@TIMBRE on 2006/01/25 19:42:31

http://bugzilla.zimbra.com/show_bug.cgi?id=5576
bug: 5576

CMB was doing the wrong thing with the -l (limit) option, it was mistakenly interpreting it to mean "last result number to show" instead of "number of results to show.

Also fixed an unrelated perf problem with CMB: it had been hardcoded for testing to only request 3 results per server roundtrip -- this was causing MANY extra round-trips when getting a large number of results.

Affected files ...

... //depot/CRAY/ZimbraNetwork/ZimbraServer/src/java/com/zimbra/cs/network/cmbsearch/CrossMailboxSearch.java#4 edit
... //depot/CRAY/ZimbraServer/src/java/com/zimbra/cs/index/ProxiedQueryResults.java#4 edit