http://bugzilla.zimbra.com/show_bug.cgi?id=5569 Bug: 5569
Optimized initial contact load as well as the
autocomplete pre-matching.

We now use the anonymous contact object
returned by the JS response, and convert it to
a ZmContact when necessary.

These changes are mainly intended to speed up
performance in IE. Loading and pre-matching 3900
contacts now takes ~41 sec, which is an improvement
of 10X - 15X.