http://bugzilla.zimbra.com/show_bug.cgi?id=5698 bug: 5698
Sync was a little bit broken.  "rev" referred to the change number for the last *content* edit; we were also using the same attribute to hold the last *metadata* edit.  For sync, return only the latter.  That attribute is now called "ms" -- metadata sequence number.

The sync client will have to be updated to accommodate this protocol change (they're still expecting "rev").