http://bugzilla.zimbra.com/show_bug.cgi?id=5971 bug: 5971
Implement redirect at the backend.
Now if you directly set your Sieve filter in LDAP to have:

if anyof (header :contains "subject" "magic" )
{
    redirect "you@gmail.com";
}

message with subject of "magic" will be resent to you@gmail.com.

If the redirect address is invalid, the message will be bounced to the sender.
If ultimately the email cannot be sent to the redirect address, the message will be
bounced to the sender.