Load testing changes:
* Fixed JavaMail javadoc link

* Moved test configuration files into ZimbraPerf/conf.

* Moved SoapGenerator and associated classes to com.zimbra.perf.soap in
ZimbraPerf.  This removes the dependency on zimbraqa.jar, so that our
braver customers can run the SOAP tests without dragging in the rest
of ZimbraQA.

* Added LoadUtil.addUsers() for defining the user set with a single
API call.

* Updated IMAP tests to use LoadUtil.addUsers(), so that mail is sent
only to users that are participating in the current run.  This makes
sure that mailboxes for other users in the user pool don't get
instantiated.