http://bugzilla.zimbra.com/show_bug.cgi?id=6067 bug: 6067
Add Mailbox.getMailboxSizes() to return a map mapping zimbra ID to approximate mailbox size for all mailboxes instantiated on the system.

Note that mailboxes are created lazily, so there may be accounts homed on this system for whom there is is not yet a mailbox and hence are not included in the return value.  Sizes are checkpointed frequently, but there is no guarantee that the approximate sizes are currently accurate.