http://bugzilla.zimbra.com/show_bug.cgi?id=175 http://bugzilla.zimbra.com/show_bug.cgi?id=5996 http://bugzilla.zimbra.com/show_bug.cgi?id=6084 Bug: 175,5996,6084
Compose toolbar now has an Options button which allows
the user to change the compose form in various ways. The
actions that are available depend on whether it's a new message,
a reply, or a forward. The tooltip for the button is different for
each case.

In order of possible appearance, they are:

Reply, Reply All (reply)
HTML, Plain Text (all)
Show CC, BCC (all)
Include options (reply, forward)

The Show CC and Show BCC menu items replace the add/remove
links. They are session-persistent. We will probably add preferences
for them.

The include options control how the original message is quoted. If
you've entered text and then switch the include option, a dialog
comes up warning you that your changes will be lost.

- fix AjxVector.addList() for use by child window
- "Smart Include" -> "Include last message only"
- Signature button now shown/hidden on settings change,
  rather than added/removed
- "New Window" button moved to right
- "Save Draft" button moved left
- ZmSetting.setValue() now takes skipNotify flag

TODO:
- need some icons
- prefs for show CC/BCC