http://bugzilla.zimbra.com/show_bug.cgi?id=402 bug: 402
If the global LDAP config zimbraMessageIdDedupeCacheSize is set to a nonzero number, we keep track of the last N messages delivered through the system.  If, during delivery, we detect that the message has already been delivered to a particular mailbox, we don't redeliver the message to the mailbox.  Note that this is a memory-only cache and hence will be reset on server restart.

Messages are identified by their Message-ID header.  This is a slight risk, as broken MUAs could potentially reuse Message-IDs, but it's a well-known risk and other companies' products have been deduping on Message-ID this for quite a while.