http://bugzilla.zimbra.com/show_bug.cgi?id=6363 Bug: 6363
(!retVal) is not the same as (retVal === false)

It's not lint!

Change 22089 broke listener notification for any control
with more than one listener. That includes the tab buttons,
which led to the tab view below not getting replaced.

Since almost no listeners return a value,
AjxEventMgr.notifListeners was exiting after the first
listener was notified. The idea behind (retVal === false)
is that a listener can explicitly return false to stop further
notifications.