http://bugzilla.zimbra.com/show_bug.cgi?id=6061 bug: 6061 - Fedora Core 4, FireFox 1.5 Send/Cancel bar broken
For some reason Linux does not play well when trying to set focus on the Iframe window. So wrapped a try/catch block around this code (catch will do nothing -- for now). This allows the rest of the code to continue executing and therefore bring up the compose view normally.