http://bugzilla.zimbra.com/show_bug.cgi?id=1995 bug 1995
added ssh remote command and control related attributes

   zimbraRemoteManagement{User,Command,PrivateKeyPath,Port}

Unless we have achieved world domination, no one is ever going
to need to change these from their global config defaults.  But
it felt very wrong to set these in local config - these are 
attributes of an another server (except private key) and it would
have to be the same value on all servers in the install.

Whatever ... I get paid by the number and length of attributes, so
there.  And it's just 4 drops in our attribute ocean.  Quit your
complaining.