http://bugzilla.zimbra.com/show_bug.cgi?id=5975 Integrated change 22822 from CRAY to main.
----------

Change 22822 by bburtin@bburtin-cray on 2006/03/17 17:06:14

http://bugzilla.zimbra.com/show_bug.cgi?id=5975
Bug 5975: Daimler server stat monitoring

Changed zimbrastats.csv logic to use log4j.  The file was mysteriously
being truncated on dogfood.  Hopefully the log4j code will be robust
enough to handle whatever weirdness was happening.

Added a ZimbraStatsAppender class, which subclasses log4j's
DailyRollingFileAppender and adds a header to the top of every
new csv file.  Currently this appender is tied to zimbrastats.csv.
We can generalize it later if we want to use it for other files.

Updated collection classes in ZimbraPerf to be type-safe.

Affected files ...

... //depot/CRAY/ZimbraConvertd/conf/log4j.properties#3 edit
... //depot/CRAY/ZimbraConvertd/conf/log4j.properties.production#3 edit
... //depot/CRAY/ZimbraServer/conf/log4j.properties#3 edit
... //depot/CRAY/ZimbraServer/conf/log4j.properties.production#5 edit
... //depot/CRAY/ZimbraServer/src/java/com/zimbra/cs/stats/ZimbraPerf.java#4 edit
... //depot/CRAY/ZimbraServer/src/java/com/zimbra/cs/stats/ZimbraStatsAppender.java#1 add