http://bugzilla.zimbra.com/show_bug.cgi?id=1363 bug: 1363
Enforce a limit on the quantity of data POSTable to the UserServlet.  Use the server config value zimbraFileUploadMaxSize to set this size.  Make sure we're using longs and not ints to hold max size.

UserServlet now supports the PUT method as a synonym for POST.

Don't use ServletRequest.getParameter() to read URI parameters ("foo=bar&baz=whop"), as it interferes with ServletRequest.getInputStream() and hence breaks POST.  Move the implementation of getURIParams from HttpRequestLine to HttpUtil, add URL decoding to it, and use it to parse UserServlet query strings.