http://bugzilla.zimbra.com/show_bug.cgi?id=6454 http://bugzilla.zimbra.com/show_bug.cgi?id=2 bug: 6454 - Try #2
Only use UTC when UTC is given - this avoids breakage w/ modfiying recurring appointments.

Theres some level of complexity involved w/ UTC and timezones so I'm sure this is not the end of the bugs we'll be seeing, but AFAIK, creating/modifying appointments with UTC should now work as expected.