http://bugzilla.zimbra.com/show_bug.cgi?id=7452 bug: 7452
Support the "checked" bit on a folder, settable via the "check" and "!check" FolderAction operations.  This is to be used for persisting the set of "active" calendars/addressbooks/etc. in the web UI.

Normalized rerturning folder flags via SOAP.  We no longer return a separate "excludeFreeBusy" attribute on each folder.  We now just return a normal "f" attribute (for flags, same as on every other item).  The allowed flags on a folder are:

   # - checked in web UI
   b - exclude from free/busy calculations
   * - "subscribed" in IMAP