http://bugzilla.zimbra.com/show_bug.cgi?id=7432 bug: 7432
Map the Zimbra "Sent" folder to "Sent Items" when the user logs in with the "/wm" username suffix indicating they're on a Windows Mobile 5 device.

In the unlikely event that the hack is invoked with the "/wm" login and the user actually has a real "Sent Items" folder, the "Sent" folder will shadow the user folder.