http://bugzilla.zimbra.com/show_bug.cgi?id=7525 bug: 7525
Allow user to choose skin.

There is a COS setting which determines whether
the user can change skins. That may get replaced
at some point with a list of available skins.

In Options/General, there is a skin dropdown. If you
choose a new skin, a dialog will pop up asking you if
you'd like to reload now. If not, the change will take
effect on next login. The skin value is maintained in a
cookie.