http://bugzilla.zimbra.com/show_bug.cgi?id=7308 bug: 7308
Zimbra hosted demo application doesn't work in Firefox trunk builds

Switch over to createElementNS to set the default namespace for <context> as well as <*Request>. Doesn't seem to work on Safari so if'd it out for that (as well as browsers without doc.createElemementNS).

Tested on Firefox (1.5.0.3 +  nightly), Safari, and IE, though I woudn't be surpised if I broke something.