http://bugzilla.zimbra.com/show_bug.cgi?id=7957 Bug: 7957
Exception handler in ZCalendar.ZCalendarBuilder.build() had a loop reading the remaining bytes off the Reader, but it was using a wrong check.  Instead of looping while Reader.ready() returns true, it should have checked for the return value of Reader.read() being -1, indicating the end of the data stream.