bug:none
really truly fix the "r-t-w deletes my MySql data dir if it is in the standard place" problem.  (typo in my checkin yesterday.  Oops)

also fix the "reset-the-world-stage1" task so that it doesn't try to start tomcat -- this is more in line with the documentation and it lets me run tomcat in the debugger directly (I don't have to go and execute "tomcat stop" manually every time).  Normal r-t-w will automatically start tomcat and do the right thing.