http://bugzilla.zimbra.com/show_bug.cgi?id=8606 bug: 8606
Better handling of message parts with garbage charsets (e.g. "X-UNKNOWN").  Now we funnel requests for part content as a String to Mime.getStringContent(), which first tries to use the specified charset and then on UnsupportedEncodingException falls back to the server's default charset.

Renamed the Mime.decodeText variant that returned a Reader, and made the other one return a String rather than passing in a StringBuffer.