http://bugzilla.zimbra.com/show_bug.cgi?id=8687 bug: 8687
FreeBusy.getFreeBusyList() was using the wrong test to determine whether a folder had "exclude from free/busy" checked.  This stemmed from confusing naming of the flag bitmasks in Flag.java, so all those bitmasks have been renamed to Flag.BITMASK_*.

Added Mailbox.getFolderList() to return a typed List of the the folders in a Mailbox in some sorted order.