http://bugzilla.zimbra.com/show_bug.cgi?id=8884 bug: 8884
Allow color to be set while creating folders, searches, and mountpoints.  Allow flags to be set while creating folders and mountpoints.  UI should specify that newly-created calendar folders and calendar mountpoints be created with the "checked" flag.

The system "Calendar" folder is now created with BITMASK_CHECKED by default.  We still need a migration script to turn this flag on for all existing users.