http://bugzilla.zimbra.com/show_bug.cgi?id=8884 bug: 8884
This migration script, migrate20060708-FlagCalendarFolder.pl, will add the "Checked" flag to the main Calendar folder (id: 10) in all existing mailboxes.  This will cause the minical and calendar view to display appointments in the main calendar by default.

Database version is now 25.  Please either reset-the-world or run this upgrade script.