http://bugzilla.zimbra.com/show_bug.cgi?id=7361 bug: 7361
Check for the magic _charset_ input field to get the character set of the posted form data.  We may be able to leverage that to not need Parag's "filename1" hack, but for now this should work properly.

Use 2047 encoded-words to encode non-ASCII attachment names.  Yeah, yeah, RFC 2231 is the standard (and we support it) and RFC 2047 is expressly not allowed in field values, but I don't think 2231 support is as widespread as support for 2047 encoded-words there.

Handle non-ASCII attachment names for inter-server upload proxying.