http://bugzilla.zimbra.com/show_bug.cgi?id=7361 bug: 7361
Set content-type (and charset=utf-8) for the top-level mail app page.

Set "accept-charset" to utf-8 on the file upload <form>, and include the magic "_charset_" hidden <input> to tell us which encoding the browser is using.

Despite all our charset guidance of all-utf-8 all the time, IE is still sending the form back in the browser's default charset when possible.  It's up to the server to look at _charset_ to figure how to interpret upload form data.