Bug: none
Don't call zmtlsctl from zmsetup.pl.  It's unnecessary because zmtlsctl is called by tomcat script on every start.  Calling it from zmsetup.pl where it currently is only causes an error message to be logged to zmsetup.log because it does "zmprov gs" on a server entry that hasn't been created yet.