http://bugzilla.zimbra.com/show_bug.cgi?id=9005 bug: 9005
When a conversation is being hard deleted, no messages can be deleted, and the cause for at least one of the messages is "permission denied" or "modify conflict", throw an appropriate exception.  If nothing's being deleted because of a TargetConstraint, that's OK -- just silently do nothing.

Short-circuit deletes in MailItem if it becomes apparent that nothing is actually being deleted.