http://bugzilla.zimbra.com/show_bug.cgi?id=9360 bug: 9360
Don't NPE during proxying when an account doesn't have zimbraMailHost set.

Admin commands shouldn't refer to DocumentHandler.proxyIfNecessary, as it won't do anything on admin commands.

Create AccountDocumentHandler, which tries to proxy to the account's mail host but falls back to the local host when the remote host is nonexistent/unreachable.