http://bugzilla.zimbra.com/show_bug.cgi?id=7961 bug 7961
mihai has messages larger than 10MB which on the cross ocean link
(the link goes quiet for a while - when it's flowing it's all good) end
up in memory first (hmm, we might need to block the thread that does
writes instead of using up memory).
,
make the write queue max size be zimbraMtaMaxMessageSize instead of
coming from local config.