http://bugzilla.zimbra.com/show_bug.cgi?id=8059 bug: 8059
Get rid of "ghost" messages in IMAP, 'cause that spooky shit freaks me out.  Instead, just catch the case where a created message has the IMAP ID of an existing message in the SELECTed folder.  Notifications to other sessions should not be screwed up by COPY ops any more.

We've changed the order of notification processing to creates/deletes/modifies to make sure we catch that create first, though in practice this is never necessary.