http://bugzilla.zimbra.com/show_bug.cgi?id=9549 Bug: 9549
Focus work, mostly for dialogs.

The keyboard mgr now maintains a stack of
default key action handlers, instead of just one.
That way, a dialog can push a default handler so
that keys that don't happen in a control can be
handled. That makes Enter (submit) and Esc (cancel)
work when focus is in an input field.

- Safety check in compose view for switching modes
- Safety check in app view mgr to handle occasional
flail where app content area ends up blank except for
the sash
- Added "NW" as keyboard shortcut for "New Wiki"