http://bugzilla.zimbra.com/show_bug.cgi?id=8210 Bug: 8210
Support for keyboard nav within an HTML
compose area. The HTML editor checks to
see if the key combo is one of its shortcuts
(which are all CTRL combos), if not it passes
the key event to the keyboard mgr.

- Occasionally after switching to HTML mode
the area won't take input. That is an existing
bug caused by a failed attempt to put the
browser in design mode.