http://bugzilla.zimbra.com/show_bug.cgi?id=9014 bug: 9014
Rather than attempt to guess which ACLs to copy onto child folders created by delegates, we copy *no* ACLs to child foldrers.  Instead, we require that delegates have the both insert and inherited read permission on the parent folder to create subfolders in it.

As a result, the child folder will use the inherited permissions from its parent, and thus the folder's creator is guaranteed to be able to read the newly-created folder.

Please look for the new "c" effective permission indicating that the user can create subfolders.  You can't *grant* "c" -- it's just a note that subfolder create will succeed because the delegate has insert and inherited read access on the folder.