http://bugzilla.zimbra.com/show_bug.cgi?id=8908 bug: 8908
Rework sync code a bit.  Break out getModifiedTags() and getModifiedFolders() from Mailbox.getModifiedItems().  SyncOperation now doesn't loop over all types individiually but instead does one call for folders, one call for tags, and one call for everything else.  getModifiedItems() now returns a Pair containing all visible modified MailItems and the ids of not-visible modified items.

You can now restrict getModifiedItems() to a set of folders.  This may not be useful.

In delegate stores, when an item not visible to the user has changed since the last sync, it is reported as "deleted".  It should now be possible for ZCO to handle items moved out of shared folders by the share owner.