http://bugzilla.zimbra.com/show_bug.cgi?id=4944 bug: 4944
When a SOAP client is not sending last-seen notification sequence numbers in its requests, don't keep old changes around after notifying.  We're currently keeping them until we hit the 100-block limit, but if the client isn't sending sequence numbers then we know the notifications are going to back up and eat memory uselessly.