http://bugzilla.zimbra.com/show_bug.cgi?id=5282 Bug: 5282
More view pre-emption cleanup.

Instead of tracking whether a pending view is
poppable, we now track whether a view is
transient. A transient view doesn't go on the
hidden stack when another view is pushed;
that means it never comes into view via a pop
of a view on top of it. Currently, transient views
are related to editing: compose, appt compose,
new contact, and page edit.