http://bugzilla.zimbra.com/show_bug.cgi?id=9985 Bug: 9985
Moving the preload of compose views to
a timed action exposed a bug in the app
view mgr, which sort of assumed that a
view would be pushed after being created.

After login, try changing the height of the
browser window (the CLV doesn't change),
or open the search builder (a total mess). The
problem was that the app view mgr thought
that appt compose was the current content.

- also removed some unused code