http://bugzilla.zimbra.com/show_bug.cgi?id=10100 bug: 10100
Use the fact that ImapFolder.mSequence is sorted on IMAP UID to get rid of the UID-to-ImapMessage mapping.  Instead, do a binary search on mSequence to find items by UID.  The ImapFolder.mUIDs HashMap has been renamed mMessageIds and now only maps Zimbra item IDs to ImapMessages; the keys are now all positive.