http://bugzilla.zimbra.com/show_bug.cgi?id=10330 bug: 10330 In DwtProgressBar which is used by the quota progress display, the following code caused the issue on the linux firefox: if(AjxEnv.isLinux)  	this._row.style.lineHeight = 13;
Remove these two line to fix the bug. If this fix causes other issues, we can disable the progress bar display on the linux firefox.


Remove these two line to fix the bug. If this fix causes other issues, we can disable the progress bar display on the linux firefox.