http://bugzilla.zimbra.com/show_bug.cgi?id=1629 bug: 1629 - Filter definition is not updated after moving folder
Now handling the case where a parent folder is moved or renamed.
Confirmed that filter rules are updated and that mail continues
to be delivered into both the parent and child folders.

Fixed Folder.updateRules() so that it doesn't set rules if they
weren't affected by a folder rename.

Fixed Mailbox.createFolder() so that it passes the OperationContext
to getFolderById().  This version of the API calls beginTransaction(),
which causes the folder and tag caches to get initialized.