http://bugzilla.zimbra.com/show_bug.cgi?id=5862 http://bugzilla.zimbra.com/show_bug.cgi?id=10615 bug: 5862, 10615
SendMsgRequest can specify a "From" address for the composed mail by adding a <e t="f" .../> element to the enclosed <m>.

When sending mail, we keep an existing From header only if it matches the mailbox owner's address or alias or if it's the canonical address (see AccountUtil.addressMatchesAccount).  Otherwise, we rewrite it to match the mailbox's owner.  All Sender headers are now stripped during mail send.