http://bugzilla.zimbra.com/show_bug.cgi?id=9277 bug: 9277
Cleanup the indexing codepaths so that they make more sense -- eliminate the 'Indexer' class and merge its functionality into 'MailboxIndex'  (soon to be renamed to just 'Index').

Add hooks to the ZimbraAnalyzer to allow extensions to register external Analyzers.  

Everything is done except for the configuration part -- next step is to wade into LDAP...