http://bugzilla.zimbra.com/show_bug.cgi?id=10652 bug: 10652
Checkpoint.

Parse and generate SENT-BY parameter in iCalendar ORGANIZER and ATTENDEE properties.

On-behalf-of meeting creation.  Invite email goes out of secretary mailbox and is saved in that mailbox's Sent folder, not boss's.  ORGANIZER is boss and SENT-BY is secretary.  From email header is set to boss and Sender and Reply-To set to secretary.  Attendees reply to secretary.  Likewise for modifying or cancelling meetings.

On-behalf-of invite acceptance.  Boss is invited to a meeting and secretary may accept on behalf of boss.  iCalendar REPLY goes out of secretary's mailbox and is saved in secretary's Sent folder, not boss's.  ATTENDEE is boss and SENT-BY is secretary.  From email header is boss, Sender and Reply-To are secreatry.

Todo1: Move invite email in boss's Inbox when secretary accepts.  This is currently not being done due to ACL issue. (Secretary has no permission on boss's Inbox and Trash folders.)

Todo2: Fix reply handling when secretary receives attendees' replies.  Boss's appointment must be updated to reflect the attendance info, but currently there is a transaction/locking issue because boss mailbox update is triggered during a message delivery transaction on secretary mailbox.