http://bugzilla.zimbra.com/show_bug.cgi?id=10652 bug: 10652
When replying to a meeting invitation on behalf of another user, use the original user's OperationContext when moving invite email to Trash folder in SendInviteReply handler.  This is a hack to get the move to succeed even though the authenticated user has no access right on Inbox and Trash folders of target mailbox.